Split cluster and service autoscaler
This is just refactoring. It should be a noop in terms of functionality.
The motivation is that the two autoscalers share very little code or
functionality.
