+++ paasta_tools/autoscaling/utils.py
@@ -1,0 +1,25 @@
+#!/usr/bin/env python
+# Copyright 2015-2016 Yelp Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from collections import defaultdict
+
+
+_autoscaling_components = defaultdict(dict)
+
+
+def register_autoscaling_component(name, method_type):
+    def outer(autoscaling_method):
+        _autoscaling_components[method_type][name] = autoscaling_method
+        return autoscaling_method
+    return outer
